Update license in a few more places
This should complete 4f07ab6.

Also, before this commit, I ran filter-branch on the repository to
remove some unnecessary stuff. Sorry to everybody who is affected by
this. Try to reclone the repository and hopefully you can easily
reapply your changes if any. If not, feel free to contact me. There
shouldn't be any reasons to perform this operation again in the
future.