Change Fortran interface for integrands
Regarding numerical integration, the first argument of the integrand
is now also an AD variable. Previously, there was a term in the second
derivative of the integral (required for acceleration) that was
computed with finite differences. Thanks to this change it is now
computed with AD. With this, finite differences are not used anywhere
in the code anymore, unless the user explicitly requests that.