Add tests for nested integrals
Add tests for a double integral where the bounds of the outer integral
are y1...y2 and those of the inner integral are x1...x2. Each of the
bounds can be an active or a constant (passive) fitting parameter. All
combinations of y1 y2 x1 x2 need to be tested.