Tidy the handling of integration bounds
Remove the separate handling of the case where both bounds are active
from the AD execution trace. The fact that the lower and upper bound
can be separately handled is sufficient.

Simplify the integration functions in numerical_integration.cpp. Most
of those can be short wrappers around the main integration function.