Remove MPI and Scalapack stubs
Use preprocessor macros such as #ifdef USE_MPI instead of defining MPI
stub functions and variables. In general, yes, the use of the
preprocessor should be minimized but the latter approach has proven to
be too error-prone because it's easy to forget to define the stubs or
remove them when not needed anymore.