Use reverse::forwards instead of constants
In most operations, values that were saved in reverse::constants are now
saved in reverse::forwards. This reduces complexity because there is one
less array to manage. That said, reverse::constants is still used during
some of the numerical integration operations. However, instead of
setting a fixed size for it, it is cleared after each return sweep and
new values are added with emplace_back. This should incur very little
overhead because this array is not used often (only when we need
derivatives of the lower and/or upper bounds of an integral).