Add function to compute (J^TJ)^-1
The inverse of J^TJ is the error covariance matrix. Multiply this by
chi2 / DOF to get the uncertainty estimates of fitting parameters
(square roots would be the standard deviations). It is computed with a
call to LMsolver::getInvJTJ().

Also, add function degreesOfFreedom which returns the number of data
points minus fitting parameters.

Also, format the test directory with clang-format.