Replace MPI with OpenMP in C++ code
This is a major update which affects much of the code. The MPI+Scalapack
support made the code unnecessarily complicated and posed some
non-trivial dependencies to the user. The new version is much more
maintainable and also easier to use. The user only needs to specify one
additional parameter:

...
solver.settings.n_threads = 8;
solver.fit();
...

The code is thread-safe meaning you can call it from an OpenMP parallel
environment. Here is an example of running four simultaneous fitting
procedures with two threads each (8 threads total):

\#pragma omp parallel num_threads(4)
    solver.settings.n_threads = 2;
    solver.fit();

Note that the previous implementation made use of MPI remote memory
access (shared memory) which means it was already limited to a single
node. Switching to OpenMP doesn't add any constraints in that sense.

Bump to version 2.0.